commit 4/4 12:30 vectores

+++ TP_1_Cascara/TP_1_Cascara/funciones.h
@@ -4,7 +4,7 @@
 void menu();
 int suma(int, int);
 int resta(int, int);
-int division(int, int);
+float division(int, int);
 int mult(int, int);
 int factorial(int);
 unsigned int verifica(void);
